added skip functionality

+++ console/helpers/FreshDatabase.go
@@ -1,12 +1,26 @@
 package console_helpers
 
 import (
+	"bufio"
 	"ch/kirari04/videocms/config"
 	"ch/kirari04/videocms/inits"
+	"fmt"
 	"os"
 )
 
 func FreshDatabase() error {
+	reader := bufio.NewReader(os.Stdin)
+
+	fmt.Println("This wil delete all your files and database.")
+	fmt.Print("Confirm with DELETEALL: ")
+	msg, err := reader.ReadString('\n')
+	if err != nil {
+		return err
+	}
+	if string(msg) != "DELETEALL" {
+		fmt.Println("Skipping")
+		return nil
+	}
 	os.Remove("./database/database.sqlite")
 	os.RemoveAll(config.ENV.FolderVideoQualitysPriv)
 	os.RemoveAll(config.ENV.FolderVideoUploadsPriv)
